app/requirements.txt
flask
flask-SQLAlchemy
PyNaCl
cryptography
apscheduler
python-dotenv
Werkzeug
gunicorn
Jinja2
MarkupSafe
itsdangerous
qrcode